make migration only run on dev

+++ db/migrate/20110621170904_remove_searchable_date_time.rb
@@ -1,7 +1,7 @@
 class RemoveSearchableDateTime < ActiveRecord::Migration
   def self.up
     # for some reason this table shows up in development schema.rb but not in staging/prod
-    drop_table :searchable_date_time
+    drop_table :searchable_date_time if Rails.env.development?
   rescue
     puts "drop_table :searchable_date_time failed (safe to ignore - KV)"
   end
